Rename internal components
